+++ context_auditor/__init__.py
@@ -5,6 +5,7 @@ from .bloat import BloatThresholds, classify_bloat, summarize_traces
 from .guard import GuardConfig, ContextGrowthGuard
 from .io import load_jsonl, write_jsonl
 from .mitigation import MitigationReport, remove_exact_duplicate_segments
+from .mitigation_eval import evaluate_mitigation_strategies, final_invocations, summarize_mitigation_rows
 from .provenance import SourceType, label_message_source, segment_messages
 from .schema import AuditTrace, Message, Segment
 from .tracer import RuntimeAuditor, TraceMetadata
@@ -24,8 +25,11 @@ __all__ = [
     "classify_bloat",
     "label_message_source",
     "load_jsonl",
+    "evaluate_mitigation_strategies",
+    "final_invocations",
     "remove_exact_duplicate_segments",
     "segment_messages",
+    "summarize_mitigation_rows",
     "summarize_traces",
     "write_jsonl",
 ]